TWO-WAY BINDING
There are three ways to express two-way binding:

1. [value]="inputContent" (input)="inputContent =$event.target.value"
2. [(ngModel)]="inputContent"
3. bindon-ngModel="inputContent"

We can use ngModel to achive two-way binding with less code (2 and 3).

The second notation start with square brackets like property binding and then parenthesis like event binding. This notation '[(ngModel)]', is the combination of property and event binding.

Third notation is similar with prefix 'bindon-ngModel', 'bindon' is the combination of 'bind' property binding and 'on' event binding.

Two-way binding is just syntactic sugar over property and event binding.

NOTE: $event give us access to the event object, from there we can get the target element that raised the event. And from there we can get the value in the input field.