Defining input/output API for a component, there are two ways:
    a) As an array of strings in the component metadata:
        inputs: ['isFavorite:is-favorite'],
        outputs: ['change: favoritChange']

    b) Using the input/output decorator:

        import {Component, Input, Output, EventEmitter} from 'angular2/core'
        @Input('is-favorite') isFavorite = true;
        @Output('favoriteChange') change = new EventEmitter();

We can use aliasing to give a diferent name for the public properties of the component. For example 'is-favorit' instead 'isFavorit'.

this.change.emit({ newValue: this.isFavorite });  -> In our click handler, we use this instruction to publish an event. We need to pass an object, that represents any aditional data we want to send with the event. Here we pass an object with the current value of 'isFavorite'.