TEMPLATE
if the template code is small enough, is better to define it in the component metadata rather than in a separate file. One thing we need to be aware of when using templateUrl is that it will cost an extra http request, but only the first time it will be downloaded and then it will be cached.