Controlling Rendering of HTML - ngIf
We can show or hide elements. When we bind to the hidden property, the element and its children will stay in the DOM, they'll just be hidden. But when you use the ngIf directive they'll be remove from the DOM if the binding expression evalue is to false.

When we use it:

[hidden] -> For small elements trees
*ngIf    -> For large element trees, because this elements can take substantial memory and computing resources, plus angular make continuos to check for changes even to invisible properties. Angular will fisical remove these elements and destroy the components to free up resources.

Exception: if we have a page where an element sub-tree may be repeatedly removed or inserted in the DOM,  then it might be better to keep this elements in the DOM and bind to the hidden property to show or hide them.