Controlling Rendering of HTML - Elvis Operator
This combination ?. is what we call Elvis Operatior. This is more fluent and convinion way to go out agains nulls in property paths. We can use it multiple times in property paths.

Example:

    <li>Assigned to:{{ task.assignee?.role?.name }}</li>

With this expression, either assignee or role is null we're gonna get the blank string, otherwise the name of the role will be render on the page.